# apis/api.yaml
openapi: 3.0.0
info:
  title: Canvara Backend API
  version: 0.11.0
servers: []
...
paths:
...
  /offers:
    get:
      summary: List all offers
      tags:
        - offers
      parameters:
        - schema:
            type: string
            enum:
              - latest
              - recommended
            default: recommended
          in: query
          name: sort
          description: 'Sorting identifier for offer list.'
        - schema:
            type: string
          in: query
          name: keyword
          description: 'Search keyword for offer.'
        - schema:
            type: string
          in: query
          name: location
          description: 'Text filter based on offerer location.'
        - schema:
            type: string
            enum:
              - open
              - underway
              - suspended
          in: query
          name: status
          description: 'Status filter for offer.'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OfferMinimal'
      operationId: list_offers
      description: ''
      security:
        - AWSCognitoCanvaraPrototype: []
    post:
      summary: Create a new offer
      tags:
        - offers
      operationId: create_offer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the offer.
                overview_text:
                  type: string
                  description: Description of the offer.
                hashtags:
                  type: array
                  items:
                    type: string
                  description: List of hashtags.
              required:
                - name
                - overview_text
  '/offers/{offer_id}':
    parameters:
      - schema:
          type: string
        name: offer_id
        in: path
        required: true
    get:
      summary: Get details of a offer
      tags:
        - offers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
      operationId: get_offer
      security:
        - AWSCognitoCanvaraPrototype: []
    put:
      summary: Update details of a offer
      tags:
        - offers
      operationId: update_offer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the offer.
                overview_text:
                  type: string
                  description: Description of the offer.
                hashtags:
                  type: array
                  items:
                    type: string
                  description: List of hashtags.
                status:
                  type: string
                  description: Status update for offer.
                  enum:
                    - suspended
                    - completed
                    - active
    delete:
      summary: Delete an offer
      tags:
        - offers
      operationId: delete_offer
      responses:
        '204':
          description: No Content
      security:
        - AWSCognitoCanvaraPrototype: []
  '/offers/{offer_id}/offer_video':
    parameters:
      - schema:
          type: string
        name: offer_id
        in: path
        required: true
        description: Unique identifier of the offer.
    put:
      summary: Create request for a offer video upload
      operationId: upload_offer_video
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUploadResponse'
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUploadRequest'
            examples:
              Example:
                value:
                  filename: offer_video.mp4
                  content_type: video/mp4
  '/offers/{offer_id}/offer_video/{upload_id}':
    parameters:
      - schema:
          type: string
        name: offer_id
        in: path
        required: true
        description: Unique identifier of the offer.
      - schema:
          type: string
        name: upload_id
        in: path
        required: true
        description: Unique identifier of the file upload.
    put:
      summary: Finalize offer video upload
      operationId: finalize_offer_video_upload
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Status of the file upload.
                  enum:
                    - uploaded
              required:
                - status
    delete:
      summary: Delete a offer video
      operationId: delete_offer_video
      responses:
        '204':
          description: No Content
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
  '/offers/{offer_id}/proposals':
    get:
      summary: List all proposals
      tags:
        - offers
      parameters:
        - schema:
            type: string
            enum:
              - all
              - active
              - selected
              - rejected
              - in_progress
              - suspended
              - completed
            default: all
          in: query
          name: filter
          description: Filter for the results.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProposalMinimal'
      operationId: list_proposals
      description: ''
      security:
        - AWSCognitoCanvaraPrototype: []
    post:
      summary: Create a new proposal
      tags:
        - offers
      operationId: create_proposal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the project.
                overview_text:
                  type: string
                  description: Description of the proposal.
              required:
                - name
                - overview_text
  '/offers/{offer_id}/bookmark':
    parameters:
      - schema:
          type: string
        name: offer_id
        in: path
        required: true
        description: Unique identifier of the offer.
    put:
      summary: Bookmark an offer
      operationId: add_offer_bookmark
      responses:
        '204':
          description: ''
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
    delete:
      summary: Remove bookmark for an offer
      operationId: delete_offer_bookmark
      responses:
        '204':
          description: No Content
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
  '/proposals/{proposal_id}':
    parameters:
      - schema:
          type: string
        name: proposal_id
        in: path
        required: true
    get:
      summary: Get details of a proposal
      tags:
        - offers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
      operationId: get_proposal
      security:
        - AWSCognitoCanvaraPrototype: []
    put:
      summary: Update details of a proposal
      tags:
        - offers
      operationId: update_proposal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name of the project.
                overview_text:
                  type: string
                  description: Description of the proposal.
                status:
                  type: string
                  description: Current status of this proposal.
                  enum:
                    - active_read
                    - selected
                    - rejected
                    - in_progress
                    - suspended
                    - completed
    delete:
      summary: Delete an proposal
      tags:
        - offers
      operationId: delete_proposal
      responses:
        '204':
          description: No Content
      security:
        - AWSCognitoCanvaraPrototype: []
  '/proposals/{proposal_id}/proposal_video':
    parameters:
      - schema:
          type: string
        name: proposal_id
        in: path
        required: true
        description: Unique identifier of the proposal.
    put:
      summary: Create request for a proposal video upload
      operationId: upload_proposal_video
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserUploadResponse'
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUploadRequest'
            examples:
              Example:
                value:
                  filename: proposal_video.mp4
                  content_type: video/mp4
  '/proposals/{proposal_id}/proposal_video/{upload_id}':
    parameters:
      - schema:
          type: string
        name: proposal_id
        in: path
        required: true
        description: Unique identifier of the proposal.
      - schema:
          type: string
        name: upload_id
        in: path
        required: true
        description: Unique identifier of the file upload.
    put:
      summary: Finalize proposal video upload
      operationId: finalize_proposal_video_upload
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: Status of the file upload.
                  enum:
                    - uploaded
              required:
                - status
    delete:
      summary: Delete a proposal video
      operationId: delete_proposal_video
      responses:
        '204':
          description: No Content
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
  '/proposals/{proposal_id}/offerer-feedback':
    parameters:
      - schema:
          type: string
        name: proposal_id
        in: path
        required: true
        description: Unique identifier of the proposal.
    put:
      summary: Add feedback for offerer
      operationId: add_offerer_feedback
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comments:
                  type: string
                  description: Feedback comment
                concerns:
                  type: string
                  description: Feedback concern
                additional_comments:
                  type: string
                  description: Feedback additional comment
  '/proposals/{proposal_id}/proposer-feedback':
    parameters:
      - schema:
          type: string
        name: proposal_id
        in: path
        required: true
        description: Unique identifier of the proposal.
    put:
      summary: Add feedback for proposer
      operationId: add_proposer_feedback
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Proposal'
      tags:
        - offers
      security:
        - AWSCognitoCanvaraPrototype: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                comments:
                  type: string
                  description: Feedback comment
                concerns:
                  type: string
                  description: Feedback concern
                additional_comments:
                  type: string
                  description: Feedback additional comment
...
components:
  schemas:
...
    Offer:
      allOf:
        - $ref: '#/components/schemas/OfferMinimal'
        - title: Offer
        - type: object
          properties:
            overview_video:
              $ref: '#/components/schemas/UserUpload'
              description: Overview video for the offer.
...
    OfferMinimal:
      title: Offer
      type: object
      properties:
        offer_id:
          type: string
          description: Unique identifier for this offer.
        offerer:
          type: object
          description: Information about the owner/offerer of the offer.
          properties:
            user_id:
              type: string
              description: Unique identifier for the user.
            name:
              type: string
              description: Name of the user.
            profile_picture_url:
              type: string
              description: The URL of user's profile picture.
            title:
              type: string
              description: Title of the user.
            pronoun:
              type: string
              description: Pronoun.
            location:
              type: string
              description: Location of the user.
            department:
              type: string
              description: Department of the user.
            email:
              type: string
              description: Email of the user.
            phone_number:
              type: string
              description: Phone of the user.
        name:
          type: string
          description: Name of this offer.
        overview_text:
          type: string
          description: Overview text for the offer.
        hashtags:
          type: array
          description: Hashtags of this offer.
          items:
            type: string
        status:
          type: string
          description: Current status of this offer.
        created_at:
          type: string
          description: 'Date and time when this offer was created, in ISO 8601 format.'
        last_updated_at:
          type: string
          description: 'Date and time when this offer was last updated, in ISO 8601 format.'
        proposal_count:
          type: number
          description: Number of proposals belongs to this offer
      required:
        - name
        - overview_text
    Proposal:
      allOf:
        - $ref: '#/components/schemas/ProposalMinimal'
        - title: Proposal
        - type: object
          properties:
            overview_video:
              $ref: '#/components/schemas/UserUpload'
              description: Overview video for the proposal.
...
    UserUploadRequest:
      title: UserUploadRequest
      type: object
      properties:
        filename:
          type: string
          description: Name of the file being uploaded.
        content_type:
          type: string
          description: Content type of the file being uploaded.
      required:
        - filename
        - content_type
    UserUploadResponse:
      title: UserUploadResponse
      type: object
      properties:
        upload_id:
          type: string
          description: Unique identifier of this upload.
        url:
          type: string
          description: Url where the file should be uploaded to.
      required:
        - upload_id
        - url
    UserUpload:
      title: UserUpload
      type: object
      properties:
        upload_id:
          type: string
          description: Unique identifier of this upload.
        url:
          type: string
          description: Url for the uploaded file.
        content_type:
          type: string
          description: Content type of the uploaded file.
      required:
        - url
        - content_type
...
  securitySchemes:
    AWSCognitoCanvaraPrototype:
      type: http
      scheme: bearer